+++ app/Seggu.Dtos/CoveragesPackDto.cs
@@ -6,7 +6,7 @@ namespace Seggu.Dtos
     [Serializable]
     public class CoveragesPackDto : KeyValueDto
     {
-        public int RiskId { get; set; }
+        public long RiskId { get; set; }
         public List<CoverageDto> Coverages { get; set; }
 
     }
